+++ stroom-cluster/stroom-cluster-lock-impl-db/src/main/java/stroom/cluster/lock/impl/db/ClusterLockServiceImpl.java
@@ -26,6 +26,7 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Map.Entry;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Supplier;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
@@ -54,6 +55,11 @@ class ClusterLockServiceImpl implements ClusterLockService {
     @Override
     public void lock(final String lockName, final Runnable runnable) {
         dbClusterLock.lock(lockName, runnable);
+    }
+
+    @Override
+    public <T> T lock(final String lockName, final Supplier<T> supplier) {
+        return dbClusterLock.lock(lockName, supplier);
     }
 
     @Override