+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#2432** : Fix DB error during volume refresh on cluster boot.
+
 * Issue **#2382** : Improved charset resolution.
 
 ## [v7.0-beta.145] - 2021-09-14